bisa juga cuk

+++ database/migrations/2025_02_27_065844_create_tikets_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('tikets', function (Blueprint $table) {
             $table->id();
             $table->string('nomor_tiket');
-            $table->unsignedBigInteger('id_user');
-            $table->unsignedBigInteger('id_event');
+            $table->integer('id_user');
+            $table->integer('id_event');
             $table->foreign('id_user')->references('id')->on('users');
             $table->foreign('id_event')->references('id')->on('events');
             $table->string('nama_event');
